@@ -2,42 +2,63 @@
   "name": "@transmute/did-key-common",
   "author": "Orie Steele",
   "module": "dist/did-key-common.esm.js",
-  "version": "0.2.1-unstable.42",
+  "version": "0.3.0-unstable.0",
   "license": "Apache-2.0",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "publishConfig": {
     "access": "public"
   },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=10"
-  },
-  "scripts": {
-    "start": "tsdx watch",
-    "build": "tsdx build",
-    "test": "tsdx test --passWithNoTests",
-    "lint": "tsdx lint src --fix",
-    "prepare": "tsdx build"
-  },
-  "devDependencies": {
-    "tsdx": "^0.13.2",
-    "tslib": "^2.0.0",
-    "typescript": "^3.9.5"
-  },
-  "dependencies": {
-    "base64url": "^3.0.1",
-    "borc": "^2.1.2",
-    "canonicalize": "^1.0.3",
-    "cbor": "^5.1.0"
-  },
   "homepage": "https://github.com/transmute-industries/did-key.js/tree/master/packages/did-key-common",
   "repository": {
     "type": "git",
     "url": "https://github.com/transmute-industries/did-key.js.git",
     "directory": "packages/did-key-common"
+  },
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=14"
+  },
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "lint": "tsdx lint src --fix",
+    "prepare": "tsdx build",
+    "size": "size-limit",
+    "analyze": "size-limit --why"
+  },
+  "size-limit": [
+    {
+      "path": "dist/did-key-common.cjs.production.min.js",
+      "limit": "64 KB"
+    },
+    {
+      "path": "dist/did-key-common.esm.js",
+      "limit": "64 KB"
+    }
+  ],
+  "devDependencies": {
+    "@size-limit/preset-small-lib": "^4.12.0",
+    "@transmute/bls12381-key-pair": "^0.6.1-unstable.36",
+    "@transmute/ed25519-key-pair": "^0.6.1-unstable.36",
+    "@transmute/secp256k1-key-pair": "^0.6.1-unstable.36",
+    "@transmute/web-crypto-key-pair": "^0.6.1-unstable.36",
+    "@transmute/x25519-key-pair": "^0.6.1-unstable.36",
+    "size-limit": "^4.12.0",
+    "ts-jest": "^27.0.3",
+    "tsdx": "^0.14.1",
+    "tslib": "^2.3.0",
+    "typescript": "^4.3.2"
+  },
+  "dependencies": {
+    "@did-core/data-model": "^0.1.1-unstable.13",
+    "@did-core/did-ld-json": "^0.1.1-unstable.13",
+    "@transmute/did-context": "^0.6.1-unstable.36",
+    "@transmute/ld-key-pair": "^0.6.1-unstable.36",
+    "@transmute/security-context": "^0.6.1-unstable.36"
   }
 }
