@@ -2,40 +2,25 @@
   "name": "@transmute/did-key-web-crypto",
   "author": "Orie Steele",
   "module": "dist/did-key-web-crypto.esm.js",
-  "version": "0.2.1-unstable.42",
+  "version": "0.3.0-unstable.0",
+  "license": "Apache-2.0",
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "publishConfig": {
+    "access": "public"
+  },
   "homepage": "https://github.com/transmute-industries/did-key.js/tree/master/packages/did-key-web-crypto",
   "repository": {
     "type": "git",
     "url": "https://github.com/transmute-industries/did-key.js.git",
     "directory": "packages/did-key-web-crypto"
   },
-  "publishConfig": {
-    "access": "public"
-  },
-  "license": "Apache-2.0",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
   "files": [
     "dist",
     "src"
   ],
   "engines": {
-    "node": ">=10"
-  },
-  "scripts": {
-    "start": "tsdx watch",
-    "build": "tsdx build",
-    "test": "tsdx test --passWithNoTests",
-    "lint": "tsdx lint src --fix",
-    "prepare": "tsdx build",
-    "size": "size-limit",
-    "analyze": "size-limit --why"
-  },
-  "prettier": {
-    "printWidth": 80,
-    "semi": true,
-    "singleQuote": true,
-    "trailingComma": "es5"
+    "node": ">=14"
   },
   "size-limit": [
     {
@@ -47,15 +32,32 @@
       "limit": "10 KB"
     }
   ],
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "lint": "tsdx lint src --fix",
+    "prepare": "tsdx build",
+    "size": "size-limit",
+    "analyze": "size-limit --why"
+  },
   "devDependencies": {
-    "@size-limit/preset-small-lib": "^4.10.0",
-    "size-limit": "^4.10.0",
-    "ts-jest": "^26.5.3",
+    "@size-limit/preset-small-lib": "^4.12.0",
+    "@transmute/credentials-context": "^0.6.1-unstable.36",
+    "@transmute/did-context": "^0.6.1-unstable.36",
+    "@transmute/jose-ld": "^0.6.1-unstable.36",
+    "@transmute/json-web-signature": "^0.7.0-unstable.0",
+    "@transmute/security-context": "^0.6.1-unstable.36",
+    "@transmute/vc.js": "^0.2.1-unstable.12",
+    "axios": "^0.21.1",
+    "size-limit": "^4.12.0",
+    "ts-jest": "^27.0.3",
     "tsdx": "^0.14.1",
-    "tslib": "^2.1.0",
-    "typescript": "^4.2.3"
+    "tslib": "^2.3.0",
+    "typescript": "^4.3.4"
   },
   "dependencies": {
-    "@transmute/web-crypto-key-pair": "^0.6.1-unstable.18"
+    "@transmute/did-key-common": "^0.3.0-unstable.0",
+    "@transmute/web-crypto-key-pair": "^0.7.0-unstable.2"
   }
 }
