@@ -2,44 +2,60 @@
   "name": "@transmute/did-key.js",
   "author": "Orie Steele",
   "module": "dist/did-key.js.esm.js",
-  "version": "0.2.1-unstable.42",
+  "version": "0.3.0-unstable.0",
   "license": "Apache-2.0",
-  "publishConfig": {
-    "access": "public"
-  },
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=10"
-  },
-  "scripts": {
-    "start": "tsdx watch",
-    "build": "tsdx build",
-    "test": "tsdx test",
-    "lint": "tsdx lint src --fix",
-    "prepare": "tsdx build"
-  },
-  "devDependencies": {
-    "@transmute/did-key-test-vectors": "^0.2.1-unstable.42",
-    "tsdx": "^0.13.2",
-    "tslib": "^2.0.0",
-    "typescript": "^3.9.5"
-  },
-  "dependencies": {
-    "@transmute/did-key-bls12381": "^0.2.1-unstable.42",
-    "@transmute/did-key-ed25519": "^0.2.1-unstable.42",
-    "@transmute/did-key-secp256k1": "^0.2.1-unstable.42",
-    "@transmute/did-key-web-crypto": "^0.2.1-unstable.42",
-    "@transmute/did-key-x25519": "^0.2.1-unstable.42"
+  "publishConfig": {
+    "access": "public"
   },
   "homepage": "https://github.com/transmute-industries/did-key.js/tree/master/packages/did-key.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/transmute-industries/did-key.js.git",
     "directory": "packages/did-key.js"
+  },
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=14"
+  },
+  "size-limit": [
+    {
+      "path": "dist/did-key.js.cjs.production.min.js",
+      "limit": "10 KB"
+    },
+    {
+      "path": "dist/did-key.js.esm.js",
+      "limit": "10 KB"
+    }
+  ],
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "lint": "tsdx lint src --fix",
+    "prepare": "tsdx build",
+    "size": "size-limit",
+    "analyze": "size-limit --why"
+  },
+  "devDependencies": {
+    "@size-limit/preset-small-lib": "^4.12.0",
+    "fs-extra": "^10.0.0",
+    "size-limit": "^4.12.0",
+    "ts-jest": "^27.0.3",
+    "tsdx": "^0.14.1",
+    "tslib": "^2.3.0",
+    "typescript": "^4.3.4"
+  },
+  "dependencies": {
+    "@transmute/did-key-bls12381": "^0.3.0-unstable.0",
+    "@transmute/did-key-ed25519": "^0.3.0-unstable.0",
+    "@transmute/did-key-secp256k1": "^0.3.0-unstable.0",
+    "@transmute/did-key-test-vectors": "^0.3.0-unstable.0",
+    "@transmute/did-key-web-crypto": "^0.3.0-unstable.0",
+    "@transmute/did-key-x25519": "^0.3.0-unstable.0"
   }
 }
